do some refactoring and make some changes to the code
