+++ app/components/LandingPageBody/Logo.js
