+++ app/components/index.js
@@ -1,0 +1,18 @@
+export CreateAccountNav from './CreateAccountNav/CreateAccountNav';
+
+export CreateAccount01Body from './CreateAccountInfoBody/CreateAccount01Body';
+export CreateAccount02Body from './CreateAccountInfoBody/CreateAccount02Body';
+export CreateAccount03Body from './CreateAccountInfoBody/CreateAccount03Body';
+export CreateAccount10Body from './CreateAccountInfoBody/CreateAccount10Body';
+
+export FirstForm  from './CreateAccountForms/firstForm';
+export SecondForm from './CreateAccountForms/secondForm';
+export ThirdForm  from './CreateAccountForms/thirdForm';
+export FourthForm from './CreateAccountForms/fourthForm';
+export FifthForm  from './CreateAccountForms/fifthForm';
+export SixthForm  from './CreateAccountForms/sixthForm';
+
+export LandingPageBody from './LandingPageBody/LandingPageBody';
+export Logo from './LandingPageBody/Logo';
+
+export Header from './Header/header';
