+++ app/containers/Home/Home.scss
@@ -1,0 +1,72 @@
+/**
+home page css
+@author kishan
+**/
+.home-page {
+  .input-group {
+    width: 100%;
+  }
+  .indicator {
+    background-color: #efefef;
+    overflow: hidden;
+    padding: 14px;
+    margin-bottom: 20px;
+  }
+  .input input {
+    background-color: #efefef;
+    border: 0 none;
+    width: 100%;
+    height: 60px;
+    margin-bottom: 20px;
+    padding: 14px;
+  }
+}
+
+/**
+header's css
+@author kishan
+***/
+.header {
+  margin-bottom: 20px;
+  .icon {
+    display: inline-block;
+    img {
+      width: 40px;
+    }
+  }
+  .burger-menu {
+    position: absolute;
+    right: 0;
+    top: 0;
+    width: 117px;
+    z-index: 9;
+  }
+  .MyMenuButton {
+    text-align: right;
+  }
+  .MyMenuButton-menu {
+    position: absolute;
+    right: 0;
+    top: 0;
+    width: 117px;
+    background-color: #1c4587;
+    color: #fff;
+    li {
+      border: 1px solid #d9d9d9;
+      &:last-child {
+        border-bottom: 0 !important;
+      }
+    }
+  }
+  .flag_icon.icon img {
+    width: 30px !important;
+  }
+  .MyMenuButton-button span {
+    color: #999999;
+    font-size: 30px;
+    z-index: 10;
+  }
+}
+.plusIcon {
+  width: 40px;
+}
