+++ app/containers/Layout/base.scss
@@ -1,0 +1,66 @@
+body, html, #app {
+    height: 100%;
+}
+
+.primary-nav {
+    display:-webkit-flex;
+    display:flex;
+    list-style-type:none;
+    padding:0;
+    justify-content:flex-end;
+}
+
+img{
+    max-width: 100%;
+}
+
+ul{
+    padding: 0px !important;
+    margin: 0px !important;
+}
+
+ul,li{
+    list-style: none;
+}
+
+.left {
+    margin-right:auto;
+}
+
+.fullBg {
+  height: 100%;
+  background-color: #cfe2f3;
+}
+
+.form-spacing {
+  margin-top: 10px;
+  margin-bottom: 10px;
+  margin-right: none;
+  margin-left: none;
+}
+
+.btn-style {
+  display: block;
+  width: 100%;
+  font-size: 110%;
+}
+
+/**theme's variables**/
+.padding14{
+    padding: 14px;
+}
+.font18{
+    font-size: 18px;
+}
+.font22{
+    font-size: 22px;
+}
+.fontGray{
+    color: #666666;
+}
+.boldFont{
+    font-weight: bold;
+}
+.radius5{
+    border-radius: 5px;
+}
