+++ app/containers/index.js
@@ -1,0 +1,5 @@
+export LandingPage from './LandingPage/LandingPage';
+export HomePage from './Home/Home';
+export CreateAccountInfo from './CreateAccountInfo/CreateAccountInfo';
+export CreateAccount from './CreateAccount/CreateAccount';
+export BaseLayout from './Layout/BaseLayout';
