--- app/pages/CreateAccount05.js
@@ -1,15 +1,0 @@
-import React from 'react';
-import { Link } from 'react-router';
-import MobileLayout from './MobileLayout';
-import CreateAccountNav from '../components/CreateAccountNav';
-import SecondForm from '../components/signUpForms/secondForm';
-export default class CreateAccount05 extends React.Component {
-  render() {
-    return (
-      <MobileLayout>
-        <CreateAccountNav routerstep="/CreateAccount04" />
-        <SecondForm />
-      </MobileLayout>
-    );
-  }
-}
