--- app/theme.js
@@ -1,39 +1,0 @@
-export const buttonStyle = {
-  width: 180,
-};
-
-export const inputStyle = {
-  maxWidth: 180,
-
-};
-
-export const logoStyle = {
-  padding: 10,
-};
-
-export const primaryNav = {
-  padding: 10,
-};
-
-export const left = {
-  padding: 10,
-};
-
-export const formSpacing = {
-  marginTop: 10,
-  marginBottom: 10,
-};
-
-export const btnStyle = {
-  display: 'block',
-  width: 100,
-};
-
-export const logoPosition = {
-  marginTop: 20,
-  marginBottom: 20,
-};
-
-export const centerPlaceholderText = document.createElement('style');
-document.body.appendChild(centerPlaceholderText);
-centerPlaceholderText.innerHTML = '::-webkit-input-placeholder {text-align: center;}';
