--- views/books/form.jade
@@ -1,35 +1,0 @@
-//- Copyright 2015-2016, Google, Inc.
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
-
-// [START form]
-extends ../base.jade
-
-block content
-  h3 #{action} book
-  form(method="POST")
-    .form-group
-      label(for="title") Title
-      input.form-control(type="text", name="title", id="title", value=book.title)
-    .form-group
-      label(for="author") Author
-      input.form-control(type="text", name="author", id="author", value=book.author)
-    .form-group
-      label(for="publishedDate") Date Published
-      input.form-control(type="text", name="publishedDate", id="publishedDate", value=book.publishedDate)
-    .form-group
-      label(for="description") Description
-      input.form-control(type="text", name="description", id="description", value=book.description)
-    .form-group
-    button.btn.btn-success(type="submit") Save
-
-// [END form]
