--- views/books/list.jade
@@ -1,37 +1,0 @@
-//- Copyright 2015-2016, Google, Inc.
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
-
-extends ../base.jade
-
-block content
-  h3 Books
-
-  a(href="/books/add", class='btn btn-success btn-sm')
-    i.glyphicon.glyphicon-plus
-    span  Add book
-
-  each book in books
-    .media
-      a(href="/books/#{book.id}")
-        .media-body
-          h4= book.title
-          p= book.author
-
-  if !books.length
-    p No books found.
-
-  if nextPageToken
-    nav
-      ul.pager
-        li
-          a(href="?pageToken=#{encodeURIComponent(nextPageToken)}") More
