--- views/books/view.jade
@@ -1,34 +1,0 @@
-//- Copyright 2015-2016, Google, Inc.
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
-
-extends ../base.jade
-
-block content
-  h3 Book
-    small
-
-  .btn-group
-    a(href="/books/#{book.id}/edit", class='btn btn-primary btn-sm')
-      i.glyphicon.glyphicon-edit
-      span  Edit book
-    a(href="/books/#{book.id}/delete", class='btn btn-danger btn-sm')
-      i.glyphicon.glyphicon-trash
-      span  Delete book
-
-  .media
-    .media-body
-      h4= book.title
-        | &nbsp;
-        small= book.publishedDate
-      h5 By #{book.author||'unknown'}
-      p= book.description
